# url-shortener-web/src/app/core/api/models/url-entity.ts
/* tslint:disable */
/* eslint-disable */
export interface UrlEntity {
  count?: number;
  createdTimestamp?: string;
  id?: null | string;
  isActive?: boolean;
  longUrl?: null | string;
  shortUrl?: null | string;
}